tests(web): address PR review - revert PushNotifications refactor, use MSW fixtures in AssetsTable test
- Revert GlobalPushNotifications utils extraction (should be a separate PR)
- Refactor AssetsTable test to use shared MSW fixtures and test builders
  instead of inline mock data
