+++ apps/web/src/components/settings/PushNotifications/__tests__/GlobalPushNotifications.test.ts
@@ -7,7 +7,7 @@ import {
   _areAllSafesSelected,
   _getTotalSignaturesRequired,
   _shouldRegisterSelectedSafes,
-  _shouldUnregisterSelectedSafes,
+  _shouldUnregsiterSelectedSafes,
   _getSafesToRegister,
   _getSafesToUnregister,
   _shouldUnregisterDevice,
@@ -436,7 +436,7 @@ describe('GlobalPushNotifications', () => {
         '4': ['0x789'],
       }
 
-      const result = _shouldUnregisterSelectedSafes(selectedSafes, currentNotifiedSafes)
+      const result = _shouldUnregsiterSelectedSafes(selectedSafes, currentNotifiedSafes)
       expect(result).toBe(true)
     })
 
@@ -450,7 +450,7 @@ describe('GlobalPushNotifications', () => {
         '1': ['0x123', '0x456'],
       }
 
-      const result = _shouldUnregisterSelectedSafes(selectedSafes, currentNotifiedSafes)
+      const result = _shouldUnregsiterSelectedSafes(selectedSafes, currentNotifiedSafes)
       expect(result).toBe(true)
     })
 
@@ -464,7 +464,7 @@ describe('GlobalPushNotifications', () => {
         '1': ['0x123'],
       }
 
-      const result = _shouldUnregisterSelectedSafes(selectedSafes, currentNotifiedSafes)
+      const result = _shouldUnregsiterSelectedSafes(selectedSafes, currentNotifiedSafes)
       expect(result).toBe(true)
     })
 
@@ -479,7 +479,7 @@ describe('GlobalPushNotifications', () => {
         '4': ['0x789'],
       }
 
-      const result = _shouldUnregisterSelectedSafes(selectedSafes, currentNotifiedSafes)
+      const result = _shouldUnregsiterSelectedSafes(selectedSafes, currentNotifiedSafes)
       expect(result).toBe(false)
     })
   })
